4.3 add bottom_sheet_menu delete playlist

+++ app/src/main/java/com/practicum/playlistmaker_1/db/dao/PlaylistDao.kt
@@ -14,7 +14,7 @@ interface PlaylistDao {
     suspend fun addPlaylist(playlistEntity: PlaylistEntity)
 
     @Query("DELETE FROM ${PlaylistEntity.TABLE_NAME} WHERE id = :id")
-    suspend fun deletePlaylist(id: Int)
+    suspend fun deletePlaylist(id: Long)
 
     @Query("SELECT * FROM ${PlaylistEntity.TABLE_NAME} ORDER BY id DESC")
     suspend fun getPlaylists(): List<PlaylistEntity>
