+++ app/src/main/java/com/practicum/playlistmaker_1/media_library/data/impl/PlaylistRepositoryImpl.kt
@@ -24,7 +24,7 @@ class   PlaylistRepositoryImpl(
             .addPlaylist(dbConvertor.mapFromPlaylistToPlaylistEntity(playlist))
     }
 
-    override suspend fun deletePlaylist(id: Int) {
+    override suspend fun deletePlaylist(id: Long) {
         appDatabase.playlistDao().deletePlaylist(id)
     }
